A lot of clean up in this commit, and I've made a freindly printout of what we are actually doing.
Also, there are a few global variables that I need to figure out how to get rid of. The bacula directories and coucdb stuff should be put in the fd.conf file, but I have not figured out how to properly use the values in there as default. More work to come.